[Android] Add public API for Browser Action in support library
Add a public API for Browser Action in Custom Tabs support library. The
API checks whether Chrome is available to handle the Browser Action
request. If so, wrap the request info in a Browser Action Intent and
send out to let Chrome handle it.

The detail design of the public API and Browser Action Intent wrapper
are here:
https://docs.google.com/a/google.com/document/d/1CkRNEsXjFnzo2X7SEZP2sluIIhhKRcNGbEWwAkJd4OI/edit?usp=sharing.

BUG=706695

Review-Url: https://codereview.chromium.org/2787723002