Support library changes to allow a callback notifying the selection of default items for Browser Actions
Changes in support library to add a PendingIntent in BrowserActionsIntent
as the callback which will be launched if a default menu item is selected.
The PendingIntent will includes the id of the chosen default menu item.

This CL also adds a BroadcastReceiver in the CCT test app to handle the
callback PendingIntent and displays the id of the chosen item.

BUG=737814

Review-Url: https://codereview.chromium.org/2961203002