Injecting package name to CCT support lib AndroidManifest and merge it into imported apps
This CL tries to achieve injecting package name variable into the
provider in the AndroidManifest of CCT support lib. The AndroidManifest
has a provider whose authorities should be inflated different package
names based on the imported apps.

This is achieved by having a build variable ${package_name} in the
AndroidManifest of CCT support lib. When imported, the AndroidManifest
is merged into app's AndroidManifest and inflate with the app's package
name.

Bug: 775798
Change-Id: I9cb804a972abd5e80e853ca1a1ac7bba4627ec61
Reviewed-on: https://chromium-review.googlesource.com/874912
Commit-Queue: Lei Tian <ltian@chromium.org>
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>