Change javadoc and function name to show the source package name might be untrusted
Browser Actions creates a fake PendingIntent to tell the package name of
sending application. However, the sending application could replace it
with a PendingIntent from different applications, which could change
the source package information. This CL updates the javadoc and function
name to explicitly tell this information is self-reported and could be
untrusted.

Bug: none
Change-Id: I1e0b408e5edd16c2a644feca223292186c712d7b