Update patch set 8
Patch Set 8:

Thanks!

A couple high-level questions, to make sure I understand things correctly:

PendingIntent as Identifier:
The identifier used here is a PendingIntent. Looking at the documentation and the implementation of PendingIntent#equals() in Android, it seems to me that two PendingIntents will be equal iff they will resolve to the same activity/service/broadcast. If this is the case, this means that a client can only ever have one session, unless it explicitly creates multiple Activities for this purpose. This also makes it more difficult for libraries to use CCT, as they may not have a suitable "target" for a PendingIntent.
What is the reasoning behind using a PendingIntent as an identifier? Could newSession provide an integer instead, for instance? This way the server could resume the session using the (int, calling application) key.

Restoring sessions and binders
I'm not entirely sure how binders would work there, have you been able to test that you can successfully restore a session?
From Binder's documentation in Android, Binder objects are only valid as long as the process that created them is alive. In this case, what would be the Binder object returned by restoreSession(), would it be a new one?

PendingSession
What is the reasoning behind having a PendingSession? Can you explain a use case for it?


Sorry if these questions sound strange, I'm out of context and would like to understand the details. I haven't looked at the code in details, but from what I've seen everything looks fine :-)

Patch-set: 8
Reviewer: Benoit L <1132034@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>