Add a method to retrieve both id and bitmap for TWA notification small icon
Depending on Android version we may need either resource id or a bitmap decoded from it
to display status bar icon of TWA notification. We also need decoded bitmap for the small
icon displayed in the content of notificaiton. Currently decoding is done insecurely on
Chrome side.

Added a method getSmallIconData that returns both an id and a bitmap decoded on the client's
side. Third party apps can override getSmallIconId, but not getSmallIconData.

Bug: 892674
Change-Id: I3c3f1ea67b4df5815a3f905a66e0fa659e128c71