Enable opening site-settings in Chrome as a manage space activity for apps hosting TWAs
Android Apps can choose an Activity that Android Settings will link to
when the user presses "Manage Data" for that app. This CL, plus the
corresponding one to Chrome, https://crrev.com/c/1329978, gives TWA
developers the option to link to Chrome's site-settings for the origin
their app is verified with.

Bug: 888958
Change-Id: Ia5c6ec2379ea91fadc086d7f685f712b88320004