Allow to send multiple trusted origins from TWA clients
Currently, if the user is able to navigate or gets redirected to
another trusted origin within a TWA, the only way to ensure that other
origin gets verified by doing CustomTabsSession#verifyRelationShip
on the client side. That would delay launching Chrome, and consequently,
showing splash screen, so this needs to change.

This CL adds an extra and a helper method so that clients can specify a
list of additional origins to be verified on Chrome side. Also, warmup
is remove as it's no longer needed.

Companion CL on Chrome side:
https://chromium-review.googlesource.com/c/chromium/src/+/1422159

Bug: 856544
Change-Id: I8f3f445ded929fa191b7420757f2f10fe1733fb9