Allow to send multiple trusted origins from TWA clients
Currently, if the user is able to navigate or gets redirected to
another trusted origin within a TWA, the only way to ensure that the other
origin gets verified is by calling CustomTabsSession#verifyRelationship
on the client side. Besides being a bit awkward api, that would delay
launching Chrome, and consequently, showing splash screen, so this needs
to change.

This CL adds an extra and a helper method so that clients can specify a
list of additional origins to be verified on Chrome side. Also, warmup
call is removed as it's no longer needed.
(see https://chromium-review.googlesource.com/c/chromium/src/+/1412452)

Companion CL on Chrome side:
https://chromium-review.googlesource.com/c/chromium/src/+/1422159

Bug: 856544
Change-Id: I8f3f445ded929fa191b7420757f2f10fe1733fb9
Reviewed-on: https://chromium-review.googlesource.com/c/1422138
Commit-Queue: Pavel Shmakov <pshmakov@chromium.org>
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>