Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Hey Sergey,
> 
> So while your code works, you've changed how LauncherActivity is meant to work. To specify the status bar colour, the metadata in your AndroidManifest.xml should point to a resource in colors.xml, eg:
> 
>     <meta-data
>         android:name="android.support.customtabs.trusted.STATUS_BAR_COLOR"
>         android:resource="@color/colorPrimary" />
> 
> I suspect you were getting the original crash because you were writing something like this:
> 
>     <meta-data
>         android:name="android.support.customtabs.trusted.STATUS_BAR_COLOR"
>         android:resource="0xff97c11f" />

Hi Peter,
That is my code and it looks like you said "it meant to work", but it fails at runtime. May be i'm missing something.
 AndroidManifest.xml:
    <meta-data
                android:name="android.support.customtabs.trusted.STATUS_BAR_COLOR"
                android:value="@color/colorPrimary" />

  colors.xml:
    <?xml version="1.0" encoding="utf-8"?>
    <resources>
        <color name="colorPrimary">#97c11f</color>
        <color name="colorPrimaryDark">#9cc527</color>
    </resources>

For me it looks like we get color int from resource "metaData.getInt" and that is enought to set color.
The second parameter of ContextCompat.getColor is a "resource ID" but we pass a color int (should we?) and there is no resource id with this color. As a result Resources$NotFoundException
https://developer.android.com/reference/android/support/v4/content/ContextCompat#getcolor

I even tried to link resource color in colors:
  <?xml version="1.0" encoding="utf-8"?>
     <resources>
        <color name="colorPrimary">@color/colorPrimaryDark</color>
        <color name="colorPrimaryDark">#9cc527</color>
  </resources>

Same result... I'm not android developer, may be i miss something obvious

Patch-set: 3
Reviewer: Gerrit User 1335607 <1335607@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>