Additional parameters for splash screens
A few additional parameters are used on the client side and passed into
Chrome:
- ScaleType. This is important e.g. for full-screen images (as opposed
to logo + background) - in this case clients would need ScaleType.FitXY
- Transformation matrix. Should be provided for ScaleType.MATRIX.
This may be handy e.g. to tune the position of the logo on the splash
screen.
- Fade-out animation duration. Allows to opt in for fade-out and to
specify the duration.

Support in Chrome:
https://chromium-review.googlesource.com/c/chromium/src/+/1552817

Bug: 856544
Change-Id: Ia36d26b191185be3927115aaa2367de87471d971
Reviewed-on: https://chromium-review.googlesource.com/c/custom-tabs-client/+/1552816
Commit-Queue: Pavel Shmakov <pshmakov@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>