Show splash screen only after enter animation completes
The bug occurs when the all the preparations for launching a TWA
(service connection etc) are finished sooner than the enter animation
completes. It seems we have no choice but to postpone displaying
the splash screen on the client side until onEnterAnimationComplete
is triggered.

Bug: 956361
Change-Id: Ie2f6f4548272d90069a35eca78532f6deefd8b0c
Reviewed-on: https://chromium-review.googlesource.com/c/custom-tabs-client/+/1584143
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Commit-Queue: Pavel Shmakov <pshmakov@chromium.org>