Update patch set 9
Patch Set 9:

> Patch Set 8: Code-Review+1
> 
> (1 comment)
> 
> Let me unblock you as I think this change definitely moves in the right direction.
> 
> But let's talk a bit more about the plans for surfacing these APIs to third party apps. If we are planning to release simple/quick APIs through support lib and then more customizations through some other library with a different release pipeline, I would strongly expect that everyone uses support lib other than a few key partners which we have convinced with direct interactions and the customizations go widely unused.
> 
> Do you mind giving a summary of how we are planning to expose these (just to make sure I am getting the plan right).

The plane is to have two libraries, a "high-level" and a "low-level" one. The first step would be to upstream the "low-level" part to androidx. The github library will effectively become the "high-level" one. This is what is currently being used, so at this stage nothing will effectively change for the developers, except the option to use only the low-level part, if that's all they need.

The next steps are not yet decided (in particular, whether the high-level part needs to go to androidx as well).

Patch-set: 9